scraper/requirements.txt
python-dotenv==0.17.0
instaloader==4.7.1
celery==5.0.5
redis==3.5.3